templates/python/requirements.txt
hypercorn
restate_sdk